corrected STL module

+++ CPP_8/ex00/include/easyfind.hpp
@@ -11,7 +11,7 @@ struct NoOcurrenceFound : std::exception {
 };
 
 template< typename T>
-typename T::iterator easyfind(T&,double);
+typename T::iterator easyfind(T&, int);
 
 #include "easyfind.tpp"
