+++ CPP_8/ex00/include/easyfind.tpp
@@ -1,7 +1,7 @@
 #include "easyfind.hpp"
 
 template< typename T>
-typename T::iterator easyfind(T& unknown,double val) {
+typename T::iterator easyfind(T& unknown, int val) {
 	typename T::iterator ret = std::find(unknown.begin(), unknown.end(), val);
 
 	if (!*ret) throw NoOcurrenceFound();
